Fix sign-conversion and conversion warnings in lang/ and modes/
Add explicit casts to silence -Wsign-conversion and -Wconversion
warnings in language syntax modules and mode files. Patterns include
char32_t↔char, QEColor→int, size_t↔int, and strtol→int casts.

+++ lang/magpie.c
@@ -177,7 +177,7 @@ static void magpie_colorize_line(QEColorizeContext *cp,
                     style = MAGPIE_STYLE_FUNCTION;
                     break;
                 }
-                if (qe_isupper(kbuf[0]) && (start == 0 || str[start-1] != '.')) {
+                if (qe_isupper((u8)kbuf[0]) && (start == 0 || str[start-1] != '.')) {
                     /* Types are capitalized */
                     style = MAGPIE_STYLE_TYPE;
                     break;
